requirements.txt
selenium
requests
python-dotenv
pytz

# Dev resources
ipdb
ipython
ipdb
